requirements.txt
streamlit
google-generativeai
chromadb
wikipedia
watchdog
